layers/shared/requirements.txt
asyncpg>=0.30.0
pyjwt[crypto]>=2.10.0
httpx>=0.28.0
tenacity>=9.0.0
boto3>=1.35.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
bcrypt>=4.0.0
